Enhance Post model with owner relationship and update post creation logic to enforce ownership checks

+++ app/router/post.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, Depends, HTTPException, status
-from sqlalchemy.orm import Session
+from sqlalchemy.orm import Session , relationship
 from typing import List
 
 from .. import models, schemas 
@@ -28,7 +28,7 @@ def get_db():
 def create_post(post: schemas.PostCreate, db: Session = Depends(get_db) , current_user :int = Depends(Oauth2.get_current_user)):
     print(current_user.email)
     new_post = models.Post(**post.dict(),
-                           owner_id = current_user.id)
+                           owner_id = current_user.id) 
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -41,7 +41,7 @@ def create_post(post: schemas.PostCreate, db: Session = Depends(get_db) , curren
 # -------------------------
 @router.get("/", response_model=List[schemas.Post])
 def get_posts(db: Session = Depends(get_db) , current_user :int = Depends(Oauth2.get_current_user)):
-    posts = db.query(models.Post).all()
+    posts = db.query(models.Post).filter(models.Post.owner_id == current_user.id).all()
     return posts
 
 
@@ -77,6 +77,10 @@ def update_post(id: int, updated_post: schemas.PostCreate, db: Session = Depends
             detail=f"Post with id {id} not found"
         )
 
+    if post.owner_id !=  current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, 
+                            detail="Not Authrozed to perofrm action")
+    
     post_query.update(updated_post.dict(), synchronize_session=False)
     db.commit()
 
@@ -97,7 +101,9 @@ def delete_post(id: int, db: Session = Depends(get_db), current_user :int = Depe
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with id {id} not found"
         )
-
+    if post.owner_id !=  current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, 
+                            detail="Not Authrozed to perofrm action")
     post_query.delete(synchronize_session=False)
     db.commit()
